added prims and kruskals algorithm (greedy)
