+++ Algorithms/Geometry Algorithms/Graph Theory/Dijkstra and Krushkal/Makefile
