+++ Algorithms/Geometry Algorithms/Graph Theory/Dijkstra and Krushkal/func.cpp
